requirements.txt
setuptools
python-docx
PyPDF2
